refactor(cli): modify create command
1. `create pipe` back to `create`
2. `create app [name]` to `app create [name]`
