+++ screenpipe-js/cli/src/index.ts
@@ -1,12 +1,13 @@
 import { Command } from "commander";
 import { 
+  appCommands,
   loginCommand,
   componentsCommands, 
   registerCommand,
   publishCommand,
   listVersionsCommand,
   logoutCommand,
-  createCommands
+  createPipeCommand
 } from "./commands";
 
 process.on("SIGINT", () => process.exit(0))
@@ -22,7 +23,8 @@ async function main() {
 
   program.addCommand(loginCommand)
   program.addCommand(logoutCommand)
-  program.addCommand(createCommands)
+  program.addCommand(appCommands)
+  program.addCommand(createPipeCommand)
   program.addCommand(componentsCommands)
   program.addCommand(registerCommand)
   program.addCommand(publishCommand)
